refactor: extract analytical kinematics package
