+++ docs/source/notebooks/3_1_experiment_smoothing_splines.md
@@ -953,8 +953,8 @@ To keep things simple for the first iteration we are going to keep speed constan
 from abc import abstractmethod
 import enum
 
-from drqp_brain.geometry import Point3D
-from drqp_brain.models import HexapodLeg
+from drqp_kinematics.geometry import Point3D
+from drqp_kinematics.models import HexapodLeg
 from plotting import GaitsVisualizer
 from scipy.interpolate import make_interp_spline
