+++ docs/source/notebooks/ps5_controller.py
@@ -19,7 +19,7 @@
 # THE SOFTWARE.
 import logging
 
-from drqp_brain.geometry import Point3D
+from drqp_kinematics.geometry import Point3D
 import numpy as np
 import pygame
