+++ docs/source/notebooks/run_hexapod.py
@@ -22,7 +22,7 @@
 
 import argparse
 
-from drqp_brain.models import HexapodModel
+from drqp_kinematics.models import HexapodModel
 from drqp_brain.walk_controller import GaitType, WalkController
 from matplotlib import pyplot as plt
 from plotting import plot_hexapod, update_hexapod_plot
