+++ packages/runtime/drqp_brain/drqp_brain/geometry/__init__.py
@@ -18,7 +18,5 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from .leg import Leg3D  # noqa: F401
-from .line import Line, Line3D  # noqa: F401
-from .point import Point, Point3D, SimplePoint3D  # noqa: F401
-from .transforms import AffineTransform  # noqa: F401
+from drqp_kinematics.geometry import AffineTransform, Leg3D, Line, Line3D  # noqa: F401
+from drqp_kinematics.geometry import Point, Point3D, SimplePoint3D  # noqa: F401
