+++ packages/runtime/drqp_brain/drqp_kinematics/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2025 Anton Matosov
+# Copyright (c) 2017-2026 Anton Matosov
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -18,4 +18,4 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from drqp_kinematics.geometry.leg import *  # noqa: F401,F403
+from .models import HexapodLeg, HexapodModel, LegModel, safe_arccos  # noqa: F401
